requirements.txt
numpy
pandas
scipy
tensorflow
seaborn
pytest
keras
pytorch